Extract shared fields in shared component
