+++ operator_ui/src/pages/Chains/New.tsx
@@ -1,29 +1,30 @@
 import React, { useState } from 'react'
 import { ApiResponse, BadRequestError } from 'utils/json-api-client'
-import Button from 'components/Button'
 import * as api from 'api'
 import { useDispatch } from 'react-redux'
 import { Chain, CreateChainRequest } from 'core/store/models'
 import BaseLink from 'components/BaseLink'
 import ErrorMessage from 'components/Notifications/DefaultError'
 import { notifyError, notifySuccess } from 'actionCreators'
+import Button from 'components/Button'
 import Content from 'components/Content'
 import {
   Card,
   CardContent,
   CardHeader,
-  Checkbox,
   CircularProgress,
-  FormControlLabel,
-  FormLabel,
   Grid,
   TextField,
 } from '@material-ui/core'
 import {
+  ChainConfigFields,
+  ConfigOverrides,
+} from 'pages/Chains/ChainConfigFields'
+import {
   createStyles,
   Theme,
-  WithStyles,
   withStyles,
+  WithStyles,
 } from '@material-ui/core/styles'
 
 const styles = (theme: Theme) =>
@@ -56,45 +57,28 @@ function apiCall({
   return api.v2.chains.createChain(definition)
 }
 
-interface ConfigOverrides {
-  [attr: string]: string | boolean
-}
-
-const defaultKeySpecifics = '{}'
-
 export const New = ({
   classes,
 }: {
   classes: WithStyles<typeof styles>['classes']
 }) => {
   const dispatch = useDispatch()
-  const [overrides, setOverrides] = useState<ConfigOverrides>({})
-  const [keySpecificOverrides, setKeySpecificOverrides] =
-    useState<string>(defaultKeySpecifics)
+
   const [chainID, setChainID] = useState<string>('')
+  const [overrides, setOverrides] = useState<ConfigOverrides>({})
+  const [chainIDErrorMsg, setChainIDErrorMsg] = useState<string>('')
   const [keySpecificOverridesErrorMsg, setKeySpecificOverridesErrorMsg] =
     useState<string>('')
-  const [chainIDErrorMsg, setChainIDErrorMsg] = useState<string>('')
   const [loading, setLoading] = useState<boolean>(false)
 
-  function validate({
-    chainID,
-    keySpecificOverrides,
-  }: {
-    chainID: string
-    keySpecificOverrides: string
-  }) {
+  function validate(chainID: string) {
     let valid = true
+
     if (!(parseInt(chainID, 10) > 0)) {
       setChainIDErrorMsg('Invalid chain ID')
       valid = false
     }
-    try {
-      JSON.parse(keySpecificOverrides)
-    } catch (e) {
-      setKeySpecificOverridesErrorMsg('Invalid key specific overrides')
-      valid = false
-    }
+
     return valid
   }
 
@@ -103,44 +87,29 @@ export const New = ({
     setChainIDErrorMsg('')
   }
 
-  function handleOverrideChange(event: React.ChangeEvent<HTMLInputElement>) {
-    const newOverrides = {
-      ...overrides,
-      [event.target.name]:
-        // Supports setting boolean values, since the checked status is not available on `target.value`
-        event.target.type == 'checkbox'
-          ? event.target.checked
-          : event.target.value,
+  async function onConfigChange(
+    config: { [attr: string]: string | boolean },
+    error: string,
+  ) {
+    if (error) {
+      setKeySpecificOverridesErrorMsg(error)
+      return
     }
 
-    setOverrides(newOverrides)
+    setOverrides(config)
   }
 
-  function handleKeySpecificOverrideChange(
-    event: React.ChangeEvent<HTMLTextAreaElement>,
-  ) {
-    setKeySpecificOverrides(event.target.value)
-    setKeySpecificOverridesErrorMsg('')
-  }
-
-  async function handleSubmit(event: React.FormEvent<HTMLFormElement>) {
+  function handleSubmit(event: React.FormEvent<HTMLFormElement>) {
     event.preventDefault()
-    const isValid = validate({ chainID, keySpecificOverrides })
+
+    const isValid = validate(chainID) && !keySpecificOverridesErrorMsg
 
     if (isValid) {
       setLoading(true)
 
-      const config: ConfigOverrides = {
-        ...overrides,
-      }
-
-      if (keySpecificOverrides != defaultKeySpecifics) {
-        config.KeySpecific = JSON.parse(keySpecificOverrides)
-      }
-
       apiCall({
-        config,
         chainID,
+        config: { ...overrides },
       })
         .then(({ data }) => {
           dispatch(notifySuccess(SuccessNotification, data))
@@ -178,399 +147,7 @@ export const New = ({
                     />
                   </Grid>
 
-                  <Grid item xs={12} style={{ marginTop: 10 }}>
-                    <FormLabel>Config Overrides</FormLabel>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <TextField
-                        label="Block History Estimator Block Delay"
-                        name="BlockHistoryEstimatorBlockDelay"
-                        placeholder="BlockHistoryEstimatorBlockDelay"
-                        type="number"
-                        fullWidth
-                        onChange={handleOverrideChange}
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <TextField
-                        label="Block History Estimator Block History Size"
-                        name="BlockHistoryEstimatorBlockHistorySize"
-                        placeholder="BlockHistoryEstimatorBlockHistorySize"
-                        type="number"
-                        fullWidth
-                        onChange={handleOverrideChange}
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <TextField
-                        label="Eth Tx Reaper Threshold"
-                        name="EthTxReaperThreshold"
-                        placeholder="EthTxReaperThreshold"
-                        type="text"
-                        fullWidth
-                        onChange={handleOverrideChange}
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <TextField
-                        label="Eth Tx Resend After Threshold"
-                        name="EthTxResendAfterThreshold"
-                        placeholder="EthTxResendAfterThreshold"
-                        type="text"
-                        fullWidth
-                        onChange={handleOverrideChange}
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <FormControlLabel
-                        control={
-                          <Checkbox
-                            name="EvmEIP1559DynamicFees"
-                            onChange={(event) => handleOverrideChange(event)}
-                          />
-                        }
-                        label="EvmEIP1559DynamicFees"
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <TextField
-                        label="Evm Finality Depth"
-                        name="EvmFinalityDepth"
-                        placeholder="EvmFinalityDepth"
-                        type="number"
-                        fullWidth
-                        onChange={handleOverrideChange}
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <TextField
-                        label="Evm Gas Bump Percent"
-                        name="EvmGasBumpPercent"
-                        placeholder="EvmGasBumpPercent"
-                        type="number"
-                        fullWidth
-                        onChange={handleOverrideChange}
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <TextField
-                        label="Evm Gas Bump Tx Depth"
-                        name="EvmGasBumpTxDepth"
-                        placeholder="EvmGasBumpTxDepth"
-                        type="number"
-                        fullWidth
-                        onChange={handleOverrideChange}
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <TextField
-                        label="Evm Gas Bump Wei"
-                        name="EvmGasBumpWei"
-                        placeholder="EvmGasBumpWei"
-                        type="number"
-                        fullWidth
-                        onChange={handleOverrideChange}
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <TextField
-                        label="Evm Gas Limit Default"
-                        name="EvmGasLimitDefault"
-                        placeholder="EvmGasLimitDefault"
-                        type="number"
-                        fullWidth
-                        onChange={handleOverrideChange}
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <TextField
-                        label="Evm Gas Limit Multiplier"
-                        name="EvmGasLimitMultiplier"
-                        placeholder="EvmGasLimitMultiplier"
-                        type="number"
-                        fullWidth
-                        onChange={handleOverrideChange}
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <TextField
-                        label="Evm Gas Price Default"
-                        name="EvmGasPriceDefault"
-                        placeholder="EvmGasPriceDefault"
-                        type="number"
-                        fullWidth
-                        onChange={handleOverrideChange}
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <TextField
-                        label="Evm Gas Tip Cap Default"
-                        name="EvmGasTipCapDefault"
-                        placeholder="EvmGasTipCapDefault"
-                        type="number"
-                        fullWidth
-                        onChange={handleOverrideChange}
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <TextField
-                        label="Evm Gas Tip Cap Minimum"
-                        name="EvmGasTipCapMinimum"
-                        placeholder="EvmGasTipCapMinimum"
-                        type="number"
-                        fullWidth
-                        onChange={handleOverrideChange}
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <TextField
-                        label="Evm Head Tracker History Depth"
-                        name="EvmHeadTrackerHistoryDepth"
-                        placeholder="EvmHeadTrackerHistoryDepth"
-                        type="number"
-                        fullWidth
-                        onChange={handleOverrideChange}
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <TextField
-                        label="Evm Head Tracker Max Buffer Size"
-                        name="EvmHeadTrackerMaxBufferSize"
-                        placeholder="EvmHeadTrackerMaxBufferSize"
-                        type="number"
-                        fullWidth
-                        onChange={handleOverrideChange}
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <TextField
-                        label="Evm Head Tracker Sampling Interval"
-                        name="EvmHeadTrackerSamplingInterval"
-                        placeholder="EvmHeadTrackerSamplingInterval"
-                        type="text"
-                        fullWidth
-                        onChange={handleOverrideChange}
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <TextField
-                        label="Evm Log Backfill Batch Size"
-                        name="EvmLogBackfillBatchSize"
-                        placeholder="EvmLogBackfillBatchSize"
-                        type="number"
-                        fullWidth
-                        onChange={handleOverrideChange}
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <TextField
-                        label="Evm Max Gas Price Wei"
-                        name="EvmMaxGasPriceWei"
-                        placeholder="EvmMaxGasPriceWei"
-                        type="number"
-                        fullWidth
-                        onChange={handleOverrideChange}
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <FormControlLabel
-                        control={
-                          <Checkbox
-                            name="EvmNonceAutoSync"
-                            // onChange={handleOverrideChange}
-                          />
-                        }
-                        label="Evm Nonce Auto Sync"
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <TextField
-                        label="Evm RPC Default Batch Size"
-                        name="EvmRPCDefaultBatchSize"
-                        placeholder="EvmRPCDefaultBatchSize"
-                        type="number"
-                        fullWidth
-                        onChange={handleOverrideChange}
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <TextField
-                        label="Flags Contract Address"
-                        name="FlagsContractAddress"
-                        placeholder="FlagsContractAddress"
-                        fullWidth
-                        onChange={handleOverrideChange}
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <TextField
-                        label="Gas Estimator Mode"
-                        name="GasEstimatorMode"
-                        placeholder="GasEstimatorMode"
-                        fullWidth
-                        onChange={handleOverrideChange}
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <TextField
-                        label="Chain Type"
-                        name="ChainType"
-                        placeholder="ChainType"
-                        fullWidth
-                        onChange={handleOverrideChange}
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <TextField
-                        label="Min Incoming Confirmations"
-                        name="MinIncomingConfirmations"
-                        placeholder="MinIncomingConfirmations"
-                        type="number"
-                        fullWidth
-                        onChange={handleOverrideChange}
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <TextField
-                        label="Min Required Outgoing Confirmations"
-                        name="MinRequiredOutgoingConfirmations"
-                        placeholder="MinRequiredOutgoingConfirmations"
-                        type="number"
-                        fullWidth
-                        onChange={handleOverrideChange}
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <TextField
-                        label="Minimum Contract Payment"
-                        name="MinimumContractPayment"
-                        placeholder="MinimumContractPayment"
-                        type="number"
-                        fullWidth
-                        onChange={handleOverrideChange}
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={6}>
-                    <Grid item xs={6}>
-                      <TextField
-                        label="OCR Observation Timeout"
-                        name="OCRObservationTimeout"
-                        placeholder="OCRObservationTimeout"
-                        type="number"
-                        fullWidth
-                        onChange={handleOverrideChange}
-                      />
-                    </Grid>
-                  </Grid>
-
-                  <Grid item xs={12} style={{ marginTop: 10 }}>
-                    <FormLabel>Key Specific Config Overrides</FormLabel>
-                  </Grid>
-
-                  <Grid item xs={12}>
-                    <TextField
-                      error={Boolean(keySpecificOverridesErrorMsg)}
-                      value={keySpecificOverrides}
-                      onChange={handleKeySpecificOverrideChange}
-                      helperText={
-                        Boolean(keySpecificOverridesErrorMsg) &&
-                        keySpecificOverridesErrorMsg
-                      }
-                      autoComplete="off"
-                      label={'JSON'}
-                      rows={10}
-                      rowsMax={25}
-                      placeholder={'Paste JSON'}
-                      multiline
-                      margin="normal"
-                      name="KeySpecific"
-                      id="chainConfig"
-                      variant="outlined"
-                      fullWidth
-                    />
-                  </Grid>
+                  <ChainConfigFields onChange={onConfigChange} />
 
                   <Grid item xs={12}>
                     <Button
